refactor github to new package structure
