+++ cmd/githubprcollector/BUILD.bazel
@@ -7,8 +7,8 @@ go_library(
     importpath = "github.com/stebennett/squad-dashboard/cmd/githubprcollector",
     visibility = ["//visibility:private"],
     deps = [
-        "//pkg/githubrepository",
-        "//pkg/githubservice",
+        "//pkg/github/repo",
+        "//pkg/github/service",
         "//vendor/github.com/Netflix/go-env",
     ],
 )
