+++ pkg/github/repo/BUILD.bazel
@@ -1,0 +1,12 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "repo",
+    srcs = ["repository.go"],
+    importpath = "github.com/stebennett/squad-dashboard/pkg/github/repo",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//pkg/github/models",
+        "//vendor/github.com/lib/pq",
+    ],
+)
