--- pkg/githubrepository/BUILD.bazel
@@ -1,12 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
-
-go_library(
-    name = "githubrepository",
-    srcs = ["githubrepository.go"],
-    importpath = "github.com/stebennett/squad-dashboard/pkg/githubrepository",
-    visibility = ["//visibility:public"],
-    deps = [
-        "//pkg/githubmodels",
-        "//vendor/github.com/lib/pq",
-    ],
-)
